+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/LegacyHandlerMapping.java
@@ -1,0 +1,11 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe;
+
+import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
+
+public interface LegacyHandlerMapping {
+
+    void initialize();
+
+    Controller getHandler(String requestURI);
+
+}
